+++ index.html
@@ -11,7 +11,7 @@
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
   <link type="text/css" rel="stylesheet" href="css/materialize.css" media="screen,projection" />
   <link type="text/css" rel="stylesheet" href="css/style.css">
-  <script src="build.js"></script>
+  <script src="dist/build.js"></script>
 </head>
 
 <body>
